requirements.txt
# Core
discord.py>=2.4.0
aiohttp>=3.9.0

# Database
SQLAlchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Captcha token signing
itsdangerous>=2.1.0

# Utilities
aiofiles>=23.2.0

# Image Processing
Pillow>=10.0.0
